Task4: receive data packet from TAP

+++ cpu/cc2538/stellaris_ether/stellaris_ether_netdev.h
@@ -32,8 +32,8 @@ typedef struct
 {
     netdev_t netdev;                    /**< netdev parent struct */
 
-    uint16_t rx_len;                     /**< number of bytes received */
-    uint16_t tx_len;                     /**< number of bytes in transmit buffer */
+    size_t rx_len;                     /**< number of bytes received */
+    size_t tx_len;                     /**< number of bytes in transmit buffer */
 
     uint8_t  rx_buf[ETHERNET_DATA_LEN];  /**< receive buffer */
     uint8_t  tx_buf[ETHERNET_DATA_LEN];  /**< transmit buffer */
